Rename pkg.
